--- src/components/calendar/index.js
@@ -1,2 +1,0 @@
-export { default as Calendar } from './view/colaborador';
-export { default as CalendarioEspecilista } from './view/especialista';
