--- src/components/calendar/view/colaborador/calendar-toolbar.jsx
@@ -1,148 +1,0 @@
-import { useState } from 'react';
-import PropTypes from 'prop-types';
-
-import Stack from '@mui/material/Stack';
-import Button from '@mui/material/Button';
-import MenuItem from '@mui/material/MenuItem';
-import IconButton from '@mui/material/IconButton';
-import Typography from '@mui/material/Typography';
-import LinearProgress from '@mui/material/LinearProgress';
-
-import { useResponsive } from 'src/hooks/use-responsive';
-
-import Iconify from 'src/components/iconify';
-import CustomPopover, { usePopover } from 'src/components/custom-popover';
-
-import LabelsDialog from './labels-dialog';
-
-// ----------------------------------------------------------------------
-
-const VIEW_OPTIONS = [
-  {
-    value: 'dayGridMonth',
-    label: 'Mes',
-    icon: 'mingcute:calendar-month-line',
-  },
-  { value: 'timeGridWeek', label: 'Semana', icon: 'mingcute:calendar-week-line' },
-  { value: 'timeGridDay', label: 'Día', icon: 'mingcute:calendar-day-line' },
-];
-
-// ----------------------------------------------------------------------
-
-export default function CalendarToolbar({
-  date,
-  view,
-  labels,
-  loading,
-  onToday,
-  onNextDate,
-  onPrevDate,
-  onChangeView,
-  onOpenFilters,
-}) {
-  const smUp = useResponsive('up', 'sm');
-
-  const popover = usePopover();
-
-  const [openLabels, setOpenLabels] = useState(false)
-
-  const selectedItem = VIEW_OPTIONS.filter((item) => item.value === view)[0];
-
-  const fechaTitulo = new Intl.DateTimeFormat('es-MX', { year: 'numeric', month: 'long' }).format(date)
-
-  return (
-    <>
-      <Stack
-        direction={{ xs : "column", md: 'row'}}
-        alignItems="center"
-        justifyContent="space-between"
-        sx={{ p: 2.5, pr: 2, position: 'relative' }}
-      >
-        {smUp && (
-          <Button
-            size="small"
-            color="inherit"
-            onClick={popover.onOpen}
-            startIcon={<Iconify icon={selectedItem.icon} />}
-            endIcon={<Iconify icon="eva:arrow-ios-downward-fill" sx={{ ml: -0.5 }} />}
-          >
-            {selectedItem.label}
-          </Button>
-        )}
-
-        <Stack direction="row" alignItems="center" spacing={1}>
-          <IconButton onClick={onPrevDate}>
-            <Iconify icon="eva:arrow-ios-back-fill" />
-          </IconButton>
-
-          <Typography variant="h6">{fechaTitulo}</Typography>
-
-          <IconButton onClick={onNextDate}>
-            <Iconify icon="eva:arrow-ios-forward-fill" />
-          </IconButton>
-        </Stack>
-
-        <Stack direction={{ xs : "column", md: 'row'}} alignItems='center' spacing={1}>
-          <Button size="small" color="error" variant="contained" onClick={onToday} fullWidth={!smUp}>
-            Hoy
-          </Button>
-          <Button size="small" color="primary" variant="contained" onClick={() => setOpenLabels(true)} fullWidth={!smUp}>
-            Leyenda
-          </Button>
-        </Stack>
-
-        <LabelsDialog
-          open={openLabels}
-          labels={labels}
-          onClose={() => setOpenLabels(false)}
-        />
-
-        {loading && (
-          <LinearProgress
-            color="inherit"
-            sx={{
-              height: 2,
-              width: 1,
-              position: 'absolute',
-              bottom: 0,
-              left: 0,
-            }}
-          />
-        )}
-      </Stack>
-
-      <CustomPopover
-        open={popover.open}
-        onClose={popover.onClose}
-        arrow="top-left"
-        sx={{ width: 160 }}
-      >
-        {VIEW_OPTIONS.map((viewOption) => (
-          <MenuItem
-            key={viewOption.value}
-            selected={viewOption.value === view}
-            onClick={() => {
-              popover.onClose();
-              onChangeView(viewOption.value);
-            }}
-          >
-            <Iconify icon={viewOption.icon} />
-            {viewOption.label}
-          </MenuItem>
-        ))}
-      </CustomPopover>
-    </>
-  );
-}
-
-CalendarToolbar.propTypes = {
-  date: PropTypes.object,
-  loading: PropTypes.bool,
-  onChangeView: PropTypes.func,
-  onNextDate: PropTypes.func,
-  onOpenFilters: PropTypes.func,
-  onPrevDate: PropTypes.func,
-  onToday: PropTypes.func,
-  view: PropTypes.oneOf(['dayGridMonth', 'timeGridWeek', 'timeGridDay', 'listWeek']),
-  labels: PropTypes.any,
-};
