--- src/components/calendar/view/colaborador/index.js
@@ -1,1 +1,0 @@
-export { default } from './calendar-view';
