--- src/components/calendar/view/especialista/index.js
@@ -1,1 +1,0 @@
-export {default} from './calendario-view'
