--- src/components/calendar/view/especialista/style.css
@@ -1,225 +1,0 @@
-.ButtonCita {
-  --black: #000000;
-  --ch-black: #ffffff;
-  --eer-black: #0152d3;
-  --night-rider: #e8d17d;
-  --white: #dedede;
-  --af-white: #f3f3f3;
-  --ch-white: #e1e1e1;
-  border-radius: 8px;
-  width: 200px;
-  height: 40px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 10px;
-  padding: 0px 15px;
-  border: none;
-  color: black;
-  position: relative;
-  cursor: pointer;
-  font-weight: 700;
-  transition-duration: .2s;
-  background-color: var(--ch-black);
-}
-
-.ButtonCita:before,
-.ButtonCita:after {
-  content: '';
-  position: absolute;
-  left: -2px;
-  top: -2px;
-  border-radius: 10px;
-  background: linear-gradient(45deg,
-      var(--ch-black), var(--eer-black),
-      var(--night-rider), var(--ch-white), var(--night-rider),
-      var(--eer-black), var(--ch-black), var(--ch-black));
-  background-size: 700%;
-  width: calc(100% + 4px);
-  height: calc(100% + 4px);
-  z-index: -1;
-  animation: steam 60s linear infinite;
-
-}
-
-.ButtonCita:hover{
-  filter: invert(100%);
-}
-
-@keyframes steam {
-  0% {
-    background-position: 0 0;
-  }
-
-  50% {
-    background-position: 400% 0;
-  }
-
-  100% {
-    background-position: 0 0;
-  }
-}
-
-
-.card .inner {
-  padding: 0px;
-  background: #222;
-  color: #fff;
-  border-radius: 8px;
-}
-
-.example-4 {
-  outline-width: 1px;
-  outline-offset: 0;
-  outline-color: rgba(0, 160, 252, 0.75);
-  outline-style: solid;
-  animation: animateOutline 2.5s ease infinite;
-  border-radius: 8px;
-}
-
-@keyframes animateOutline {
-  0% {
-    outline-width: 1px;
-    outline-offset: 0;
-    outline-color: rgba(0, 130, 206, 0);
-  }
-
-  10% {
-    outline-color: rgba(0, 130, 206, 0.75);
-  }
-
-  50% {
-    outline-width: 7px;
-    outline-offset: 4px;
-    outline-color: rgba(0, 130, 206, 0);
-  }
-
-  100% {
-    outline-width: 7px;
-    outline-offset: 4px;
-    outline-color: rgba(102, 102, 102, 0);
-  }
-  
-}
-
-@keyframes glow {
-  0% {
-    background-position: 0 0;
-  }
-
-  50% {
-    background-position: 100% 0;
-  }
-
-  100% {
-    background-position: 0 0;
-  }
-}
-
-.animate {
-  transform: scale(1.1);
-  transition: transform 0.1s ease-in-out;
-}
-
-
-.icons {
-  background-color: #ffffffdf;
-}
-
-.buttonActions {
-  
-  transition-duration: .3s;
-}
-
-
-.buttonActions:hover .bell {
-  animation: bellRing 0.9s both;
-}
-
-/* bell ringing animation keyframes*/
-@keyframes bellRing {
-  0%,
-  100% {
-    transform-origin: top;
-  }
-
-  15% {
-    transform: rotateZ(10deg);
-  }
-
-  30% {
-    transform: rotateZ(-10deg);
-  }
-
-  45% {
-    transform: rotateZ(5deg);
-  }
-
-  60% {
-    transform: rotateZ(-5deg);
-  }
-
-  75% {
-    transform: rotateZ(2deg);
-  }
-}
-
-.buttonActions:active {
-  transform: scale(0.8);
-}
-
-
-.notificar {
-  padding: 0;
-  margin: 0;
-  border: none;
-  background: none;
-  cursor: pointer;
-}
-
-.notificar {
-  --primary-color: #111;
-  --hovered-color: #c84747;
-  position: relative;
-  display: flex;
-  gap: 0.5rem;
-  align-items: center;
-}
-
-.notificar p {
-  margin: 0;
-  position: relative;
-  font-size: 20px;
-  color: var(--primary-color);
-}
-
-.notificar::after {
-  position: absolute;
-  content: "";
-  width: 0;
-  left: 0;
-  bottom: -7px;
-  background: var(--hovered-color);
-  height: 2px;
-  transition: 0.3s ease-out;
-}
-
-.notificar p::before {
-  position: absolute;
-  /*   box-sizing: border-box; */
-  content: "Subscribe";
-  width: 0%;
-  inset: 0;
-  color: var(--hovered-color);
-  overflow: hidden;
-  transition: 0.3s ease-out;
-}
-
-.notificar:hover::after {
-  width: 100%;
-}
-
-.notificar:hover p::before {
-  width: 100%;
-}
-
