+++ src/sections/calendar/index.js
@@ -1,0 +1,1 @@
+export { default as CalendarView } from './view';
