+++ src/sections/calendar/specialist/dialogs/end-appointment-dialog.jsx
@@ -37,7 +37,7 @@ import {
 
 import Iconify from 'src/components/iconify';
 
-export default function FinalizarCitaDialog() {
+export default function EndAppointmentDialog() {
   const { user } = useAuthContext();
   const [open, setOpen] = useState(true);
   const [open2, setOpen2] = useState(false);
