+++ src/sections/calendar/specialist/index.js
@@ -1,0 +1,1 @@
+export {default} from './calendar-especialist-view'
