--- src/sections/calendario/index.js
@@ -1,1 +1,0 @@
-export { default as CalendarioView } from './view';
